requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
distlib==0.3.9
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg2==1.19.4
et_xmlfile==2.0.0
filelock==3.17.0
Flask==3.0.3
fonttools==4.55.2
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.3
mysqlclient==2.2.7
numpy==2.0.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.0
pillow==10.3.0
platformdirs==4.3.6
pycryptodome==3.20.0
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
whitenoise==6.6.0
